Big changes and enhancements ! refactoring, renaming of clientID to applicationID Bugfix in reading New "TrackingID" attribute to monitor the commands sent to the physical module Input Connections SPI Factory

+++ opendevice-core/src/main/java/br/com/criativasoft/opendevice/core/dao/DeviceDao.java
@@ -21,5 +21,5 @@ import br.com.criativasoft.opendevice.core.model.Device;
  */
 public interface DeviceDao extends Dao<Device> {
 
-    public Device getByUID(long uid);
+    public Device getByUID(int uid);
 }
